+++ DlgOptions.h
@@ -1,7 +1,7 @@
 /////////////////////////////////////////////////////////////////////////////
 //                                                                         //
 //  NppSnippets - Code Snippets plugin for Notepad++                       //
-//  Copyright (C) 2010-2020 Frank Fesevur                                  //
+//  Copyright (C) 2010-2021 Frank Fesevur                                  //
 //                                                                         //
 //  This program is free software; you can redistribute it and/or modify   //
 //  it under the terms of the GNU General Public License as published by   //
@@ -21,36 +21,4 @@
 
 #pragma once
 
-#include "NppOptions.h"
-
-/////////////////////////////////////////////////////////////////////////////
-//
-
-class Options : public NppOptions
-{
-public:
-	Options() noexcept;
-	Options(std::wstring filename);
-	~Options();
-
-	bool GetToolbarIcon() noexcept			{ return _toolbarIcon; };
-	bool GetIndentSnippet() noexcept		{ return _indentSnippet; };
-	bool GetShowConsoleDlg() noexcept		{ return _showConsoleDlg; };
-	std::wstring GetPrevVersion()			{ return _prevVersion; };
-	std::wstring GetDBFile()				{ return _DBFile; };
-
-	void SetToolbarIcon(bool b) noexcept	{ _toolbarIcon = b; };
-	void SetIndentSnippet(bool b) noexcept	{ _indentSnippet = b; };
-	void SetShowConsoleDlg(bool b) noexcept	{ _showConsoleDlg = b; };
-	void SetDBFile(std::wstring s)			{ _DBFile = s; };
-
-	void Write();
-	void Read();
-
-private:
-	std::wstring _prevVersion;
-	std::wstring _DBFile;
-	bool _toolbarIcon;
-	bool _indentSnippet;
-	bool _showConsoleDlg;
-};
+extern void ShowOptionsDlg();